# docs/openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: TonForge Agent API
  version: "1.0.0"
  summary: Machine-facing API for AI agents to automate a seller's product lifecycle.
  description: |
    The **Agent API** lets an autonomous agent (or any HTTP client acting on a
    seller's behalf) manage that seller's marketplace listings, distribution
    manifests, and read their orders — without ever touching a browser session.

    ## Authentication
    Every request authenticates with a **Personal Access Token (PAT)** issued by
    a *verified* seller. Tokens look like `tfa_XXXXXXXX…` and are passed either as:

    - `Authorization: Bearer tfa_…`  (preferred, standard), or
    - `X-Agent-Token: tfa_…`         (fallback for clients that reserve `Authorization`).

    The acting wallet is read **from the token**, never from the request body or
    a header. An agent therefore cannot act for a seller other than the one who
    issued its token, no matter what it puts in the payload.

    Tokens are issued, listed, and revoked by the human seller through the
    session-authenticated `agent-tokens` endpoints (see the `Token management`
    tag). The plaintext token is shown **exactly once** at issue time.

    ## Authorization (scopes)
    A token carries a set of scopes. `listings:write` implies `listings:read`;
    `distribution:write` implies `listings:read`. A request missing a required
    scope returns `403 SCOPE_FORBIDDEN`.

    ## Hard gates (defence in depth)
...
    call time:
...
    own progress.

...
servers:
...
security:
  - agentBearer: []
  - agentHeader: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List the seller's orders
      description: >-
        Returns orders across all of the calling wallet's listings, newest first.
        Requires `orders:read`.
      security:
        - agentBearer: [orders:read]
        - agentHeader: [orders:read]
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100, maximum: 500, minimum: 1 }
      responses:
        "200":
          description: The seller's orders.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /buyer/orders:
    post:
      tags: [Agentic buying]
      operationId: buyerCreateOrder
      summary: "Buy step 1: create an order (agent pays with its own wallet)"
      description: >-
        Requires a buyer token (scope orders:buy); the buying wallet is ALWAYS the
        token's wallet. Sanctions + AML screen the paying wallet. The response's
        `payment` object is machine-actionable: send EXACTLY `amountNanoton` to
        `payToAddress` with BOTH `stateInitBase64` and `payloadBase64` attached
        (a plain comment transfer will not fund the escrow), e.g. via a TON
        Agentic Wallet / @ton/mcp — then call confirm.
      security:
        - agentToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId]
              properties:
                listingId: { type: string }
      responses:
        "200":
          description: Order created; includes escrow + payment instructions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      orderId: { type: string }
                      amountRaw: { type: string }
                      escrow:
                        type: object
                        properties:
                          address: { type: string }
                          stateInit: { type: string, description: base64 BOC — must be attached to the payment }
                          payload: { type: string, description: base64 BOC — must be attached to the payment }
                          totalAmountRaw: { type: string }
                      payment:
                        type: object
                        description: Machine-actionable payment instructions for the agent's wallet tooling.
                        properties:
                          network: { type: string, enum: [mainnet, testnet] }
                          payFromWallet: { type: string }
                          payToAddress: { type: string }
                          amountNanoton: { type: string }
                          stateInitBase64: { type: string }
                          payloadBase64: { type: string }
                          note: { type: string }
        "400": { description: Listing inactive / no collection. }
        "403": { description: Missing orders:buy scope. }
        "451": { description: Paying wallet sanctioned or AML high-risk. }
        "503": { description: Escrow compute failed — retry shortly. }
  /buyer/orders/{id}/confirm:
    post:
      tags: [Agentic buying]
      operationId: buyerConfirmOrder
      summary: "Buy step 2: verify the escrow payment on-chain"
      description: >-
        Verifies the payment landed AND the escrow reached FUNDED on-chain, then
        queues the license NFT mint (mintPending true; order stays pending_payment
        until the mint worker finishes). Idempotent — safe to retry while funds
        propagate.
      security:
        - agentToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Payment verified (or order already processed). }
        "400": { description: PAYMENT_VERIFY_FAILED — payment not visible yet. }
        "403": { description: Order belongs to a different wallet. }
        "409": { description: ESCROW_NOT_FUNDED — funded amount not confirmed on-chain yet. }
  /buyer/orders/{id}:
    get:
      tags: [Agentic buying]
      operationId: buyerGetOrder
      summary: Poll your own order
      description: >-
        State (pending_payment → paid → fulfilled), escrow + license NFT addresses,
        and the delivery payload once paid. Only the buyer token's own orders.
      security:
        - agentToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Order view. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /buyer/listings/{id}/download:
    get:
      tags: [Agentic buying]
      operationId: buyerDownload
      summary: "Buy step 3: signed download URL for a purchased listing"
      description: >-
        Returns a short-lived signed URL plus the expected sha256. Gated on the
        minted license NFT for the buyer wallet and a clean antivirus verdict;
        ≤20 URLs/day per purchase. Verify the sha256 after download.
      security:
        - agentToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Signed URL issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      url: { type: string }
                      expiresInSec: { type: integer }
                      sha256: { type: string }
        "403": { description: No entitlement / license not minted yet. }
        "404": { description: Listing or verified build not found. }
        "429": { description: Download rate limit (20/day). }
components:
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: tfa
      description: "`Authorization: Bearer tfa_…`"
    agentHeader:
      type: apiKey
      in: header
      name: X-Agent-Token
      description: "`X-Agent-Token: tfa_…` — fallback channel."
    sessionJwt:
      type: http
      scheme: bearer
      description: Seller user-session JWT. Used only by the Token management endpoints.
...
  schemas:
...
    Order:
      type: object
      properties:
        id: { type: string }
        listingId: { type: string }
        listingTitle: { type: [string, "null"] }
        buyerWallet: { type: string }
        state: { type: string }
        amountRaw: { type: string, description: Amount in nanotons (integer string). }
        currency: { type: string, example: TON }
        memo: { type: string }
        tonTxHash: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
...
    Error:
      type: object
      description: Route-level error shape.
      properties:
        error: { type: string }
        code: { type: string }
    AuthError:
      type: object
      description: Auth-middleware error shape.
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
          examples:
            noToken: { value: { success: false, message: "Agent token required", code: "NO_AGENT_TOKEN" } }
            badToken: { value: { success: false, message: "Invalid or expired token", code: "BAD_AGENT_TOKEN" } }
    Forbidden:
      description: Token lacks a required scope, or KYC is not approved.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
          examples:
            scope: { value: { success: false, message: "Missing required scope: listings:write", code: "SCOPE_FORBIDDEN" } }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Per-token rate limit exceeded.
      headers:
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer }, description: Unix epoch seconds. }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
          example: { success: false, message: "Rate limit exceeded (600/15min)", code: "RATE_LIMITED" }
...